+++ locales/zh_tw.json
@@ -29,10 +29,10 @@
     "error.prompt.timeout": "執行指令時發生錯誤：\n${detail}\n此問題可能因機器人未及時處理訊息導致，若重複發生，請將此訊息回報給機器人開發人員。",
     "example": "你好世界！",
     "failed": "失敗。",
-    "game.help.stop": "停止目前的遊戲。",
-    "game.message.running": "目前此遊戲正在進行。",
+    "game.help.stop": "停止當前的遊戲。",
+    "game.message.running": "當前此遊戲正在進行。",
     "game.message.stop": "已停止。",
-    "game.message.stop.none": "目前此遊戲沒有在進行。",
+    "game.message.stop.none": "當前此遊戲沒有在進行。",
     "help.option.legacy": "強制使用文本模式。",
     "i18n.unit.1": "萬",
     "i18n.unit.2": "億",
@@ -58,11 +58,11 @@
     "parser.admin.module.permission.denied": "「${module}」模組下的指令僅能由該對話的管理員所使用，請聯絡管理員執行此指令。",
     "parser.admin.submodule.permission.denied": "此指令僅能由該對話的管理員所使用，請聯絡管理員執行此指令。",
     "parser.command.format.invalid": "語法錯誤。\n使用「${prefix}help ${module}」檢視說明。",
-    "parser.command.running.prompt": "目前有指令正在執行，請稍後再試。",
+    "parser.command.running.prompt": "當前有指令正在執行，請稍後再試。",
     "parser.command.running.prompt2": "先前的指令正在執行中。",
     "parser.module.desc": "描述：\n${desc}",
     "parser.module.disabled.prompt": "「${module}」模組未啟用，請使用「${prefix}enable ${module}」啟用此模組。",
-    "parser.module.unloaded": "「${module}」模組目前未載入，請聯絡開發人員處理。",
+    "parser.module.unloaded": "「${module}」模組當前未載入，請聯絡開發人員處理。",
     "parser.superuser.permission.denied": "你沒有使用此指令的權限。",
     "petal.message.cost": "（使用了 ${amount} 片花瓣）",
     "petal.message.gained.limit": "但是，今日取得花瓣已達到每日上限，請明日再試。",
