+++ modules/arcaea/locales/zh_cn.json
@@ -5,6 +5,6 @@
     "arcaea.help.random": "随机一首歌曲。",
     "arcaea.help.rank.free": "查看当前免费包游玩排行。",
     "arcaea.help.rank.paid": "查看当前付费包游玩排行。",
-    "arcaea.message.download": "目前的最新版本为 ${version}。\n下载地址：${url}",
+    "arcaea.message.download": "当前的最新版本为 ${version}。\n下载地址：${url}",
     "arcaea.message.get_failed": "获取失败。"
 }
