+++ modules/core/locales/en_us.json
@@ -93,30 +93,33 @@
     "core.message.forward_msg.disable": "Forward message disabled.",
     "core.message.forward_msg.disabled": "Forward message has been disabled.",
     "core.message.forward_msg.enable": "Forward message enabled.",
-    "core.message.help.author.type1": "Module author: ",
-    "core.message.help.author.type2": "Module Author",
+    "core.message.help.author": "Module author: ",
+    "core.message.help.command_count": "${count} Command(s)",
     "core.message.help.helpdoc.address": "Online module document: \n${url}",
     "core.message.help.legacy.availables": "Currently available module(s):",
     "core.message.help.legacy.availables.none": "There are currently no modules available.",
-    "core.message.help.legacy.base": "Base Command: ",
-    "core.message.help.legacy.external": "Module Extension Command: ",
+    "core.message.help.legacy.base": "Base Modules: ",
+    "core.message.help.legacy.external": "External Modules: ",
     "core.message.help.legacy.more_information": "Use \"${prefix}help <module name>\" to view details.\nUse \"${prefix}module list\" to view all available modules.",
     "core.message.help.more_information": "Here are the help documents shown only show the enabled module(s). If you need to view the help documents of all modules, please use \"${prefix}module list\".",
     "core.message.help.more_information.document": "You can also consult the web documentation for help:\n${url}",
     "core.message.help.more_information.donate": "If you have sufficient funds, welcome to sponsor us: \n${url}",
     "core.message.help.not_found": "This module may not exist. Please check your input.",
+    "core.message.help.prompt": "Oranges are basic modules, blues are external modules, greys are disabled modules. \nModules marked in red can only be used by the admins of the session. \nUse \"${prefix}help <module name>\" to view details.",
     "core.message.help.regex.detail": "(${msg})",
     "core.message.help.regex.no_information": " (No description)",
+    "core.message.help.regex_count": "${count} Regex(es)",
     "core.message.help.support_regex": "This module supports regular expressions. The message will match with the following: ",
     "core.message.help.table.header.alias": "Command Alias",
+    "core.message.help.table.header.author": "Module Author",
     "core.message.help.table.header.base": "Base Module List",
-    "core.message.help.table.header.external": "Extended Module List",
+    "core.message.help.table.header.external": "External Module List",
     "core.message.help.table.header.help": "Help",
     "core.message.leave.confirm": "This operation is irreversible. Are you sure?",
     "core.message.leave.success": "Goodbye.",
-    "core.message.locale": "Current language: ${lang}",
     "core.message.locale.contribute": "Help with translation:\n${url}",
     "core.message.locale.langlist": "Supported language codes: ${langlist}",
+    "core.message.locale.prompt": "Current language: ${lang}",
     "core.message.locale.reload.failed": "Failed to reload the following string:\n${detail}",
     "core.message.locale.set.invalid": "Invalid language code.",
     "core.message.locale.set.prompt": "Use \"${prefix}locale <language code>\" to set the language.",
