+++ modules/core/locales/zh_cn.json
@@ -93,33 +93,36 @@
     "core.message.forward_msg.disable": "已关闭转发消息。",
     "core.message.forward_msg.disabled": "转发消息已关闭。",
     "core.message.forward_msg.enable": "已开启转发消息。",
-    "core.message.help.author.type1": "模块作者：",
-    "core.message.help.author.type2": "模块作者",
+    "core.message.help.author": "模块作者：",
+    "core.message.help.command_count": "${count} 条命令",
     "core.message.help.helpdoc.address": "在线模块文档：\n${url}",
     "core.message.help.legacy.availables": "当前可用的模块有：",
     "core.message.help.legacy.availables.none": "当前没有可用的模块。",
-    "core.message.help.legacy.base": "基础命令：",
-    "core.message.help.legacy.external": "模块扩展命令：",
-    "core.message.help.legacy.more_information": "使用“${prefix}help <对应模块名>”查看详细信息。\n使用“${prefix}module list”查看所有的可用模块。",
+    "core.message.help.legacy.base": "基础模块：",
+    "core.message.help.legacy.external": "扩展模块：",
+    "core.message.help.legacy.more_information": "使用“${prefix}help <模块名称>”查看详细信息。\n使用“${prefix}module list”查看所有的可用模块。",
     "core.message.help.more_information": "此处展示的帮助文档仅展示已开启的模块，若需要查看全部模块的帮助文档，请使用“${prefix}module list”。",
     "core.message.help.more_information.document": "你也可以通过查阅在线文档获取帮助：\n${url}",
     "core.message.help.more_information.donate": "若你有经济实力，欢迎给孩子们在爱发电上打钱：\n${url}",
     "core.message.help.not_found": "此模块可能不存在，请检查输入。",
+    "core.message.help.prompt": "橙色为基础模块，蓝色为扩展模块，灰色为未开启模块。\n标为红字的模块仅能被该对话的管理员所使用。\n使用“${prefix}help <模块名称>”查看详细信息。",
     "core.message.help.regex.detail": "（${msg}）",
     "core.message.help.regex.no_information": "（无描述）",
+    "core.message.help.regex_count": "${count} 条表达式",
     "core.message.help.support_regex": "此模块支持正则表达式，消息发送时将会匹配以下内容：",
     "core.message.help.table.header.alias": "命令别名",
+    "core.message.help.table.header.author": "模块作者",
     "core.message.help.table.header.base": "基础模块列表",
     "core.message.help.table.header.external": "扩展模块列表",
     "core.message.help.table.header.help": "帮助信息",
     "core.message.leave.confirm": "此操作不可逆，你确定吗？",
     "core.message.leave.success": "已执行，再见。",
-    "core.message.locale": "当前使用的语言：${lang}",
     "core.message.locale.contribute": "欢迎前往协助翻译：\n${url}",
     "core.message.locale.langlist": "支持的语言代码：${langlist}",
+    "core.message.locale.prompt": "当前使用的语言：${lang}",
     "core.message.locale.reload.failed": "以下字符串重载失败：\n${detail}",
     "core.message.locale.set.invalid": "语言代码无效。",
-    "core.message.locale.set.prompt": "可使用“${prefix}locale <语言代码>”设置语言。",
+    "core.message.locale.set.prompt": "使用“${prefix}locale <语言代码>”设置语言。",
     "core.message.module.disable.already": "失败：“${module}”模块已关闭。",
     "core.message.module.disable.base": "失败：“${module}”为基础模块，无法关闭。",
     "core.message.module.disable.not_found": "失败：“${module}”模块不存在。",
@@ -129,7 +132,7 @@
     "core.message.module.enable.not_found": "失败：“${module}”模块不存在。",
     "core.message.module.enable.qqchannel_global.success": "成功：为所有文字频道开启“${module}”模块。",
     "core.message.module.enable.success": "成功：开启模块“${module}”。",
-    "core.message.module.list.prompt": "使用“${prefix}help <对应模块名>”查看详细信息。",
+    "core.message.module.list.prompt": "使用“${prefix}help <模块名称>”查看详细信息。",
     "core.message.module.load.failed": "加载模块失败。",
     "core.message.module.load.not_found": "失败：该模块已加载或不存在。",
     "core.message.module.load.success": "成功加载模块：${module}",
