+++ modules/core/locales/zh_tw.json
@@ -55,13 +55,13 @@
     "core.message.admin.add.success": "成功：已將 ${user} 設定為機器人管理員。",
     "core.message.admin.already": "此成員已是機器人管理員。",
     "core.message.admin.ban.already": "此成員已被限制使用機器人。",
-    "core.message.admin.ban.list": "目前被設定限制使用機器人的成員：\n",
-    "core.message.admin.ban.list.none": "目前沒有被限制使用機器人的成員。",
+    "core.message.admin.ban.list": "當前被設定限制使用機器人的成員：\n",
+    "core.message.admin.ban.list.none": "當前沒有被限制使用機器人的成員。",
     "core.message.admin.ban.not_yet": "此成員沒有被限制使用機器人。",
     "core.message.admin.ban.self": "你不可以對自己進行此操作！",
     "core.message.admin.invalid": "ID 格式錯誤，格式應為「${sender}|<使用者 ID>」，請使用「${prefix}whoami」指令檢視使用者 ID。",
-    "core.message.admin.list": "目前被手動設定的機器人管理員：\n",
-    "core.message.admin.list.none": "目前沒有手動設定的機器人管理員。",
+    "core.message.admin.list": "當前被手動設定的機器人管理員：\n",
+    "core.message.admin.list.none": "當前沒有手動設定的機器人管理員。",
     "core.message.admin.remove.confirm": "你正在移除你自己的權限！你確定嗎？",
     "core.message.admin.remove.success": "成功：已將 ${user} 移出機器人管理員。",
     "core.message.alias.add.already": "「${alias}」別名已存在於自訂別名列表。",
@@ -93,33 +93,36 @@
     "core.message.forward_msg.disable": "已停用轉發訊息。",
     "core.message.forward_msg.disabled": "轉發訊息已停用。",
     "core.message.forward_msg.enable": "已啟用轉發訊息。",
-    "core.message.help.author.type1": "模組作者：",
-    "core.message.help.author.type2": "模組作者",
+    "core.message.help.author": "模組作者：",
+    "core.message.help.command_count": "${count} 條指令",
     "core.message.help.helpdoc.address": "線上模組說明文件：\n${url}",
-    "core.message.help.legacy.availables": "目前可用的模組有：",
-    "core.message.help.legacy.availables.none": "目前沒有可用的模組。",
-    "core.message.help.legacy.base": "基礎指令：",
-    "core.message.help.legacy.external": "模組擴充指令：",
-    "core.message.help.legacy.more_information": "使用「${prefix}help <對應模組名>」檢視詳細資訊。\n使用「${prefix}module list」檢視所有的可用模組。",
+    "core.message.help.legacy.availables": "當前可用的模組有：",
+    "core.message.help.legacy.availables.none": "當前沒有可用的模組。",
+    "core.message.help.legacy.base": "基礎模組：",
+    "core.message.help.legacy.external": "擴充模組：",
+    "core.message.help.legacy.more_information": "使用「${prefix}help <模組名稱>」檢視詳細資訊。\n使用「${prefix}module list」檢視所有的可用模組。",
     "core.message.help.more_information": "此處展示的說明文件僅展示已啟用的模組，若需要檢視全部模組的說明文件，請使用「${prefix}module list」。",
     "core.message.help.more_information.document": "你也可以透過查閱線上文件取得說明：\n${url}",
     "core.message.help.more_information.donate": "若你有經濟實力，歡迎在愛發電提供贊助：\n${url}",
     "core.message.help.not_found": "此模組可能不存在，請校對輸入。",
+    "core.message.help.prompt": "橘色為基礎模組，藍色為擴充模組，灰色為未啟用模組。\n標為紅字的模組只能被該對話的管理員所使用。\n使用「${prefix}help <模組名稱>」檢視詳細資訊。",
     "core.message.help.regex.detail": "（${msg}）",
     "core.message.help.regex.no_information": "（無描述）",
+    "core.message.help.regex_count": "${count} 條表示式",
     "core.message.help.support_regex": "此模組支援正規表示式，訊息傳送時將會尋找相符的以下內容：",
     "core.message.help.table.header.alias": "指令別名",
+    "core.message.help.table.header.author": "模組作者",
     "core.message.help.table.header.base": "基礎模組列表",
     "core.message.help.table.header.external": "擴充模組列表",
     "core.message.help.table.header.help": "說明資訊",
     "core.message.leave.confirm": "此行為不可逆，你確定嗎？",
     "core.message.leave.success": "已執行，再見。",
-    "core.message.locale": "目前使用的語言：${lang}",
     "core.message.locale.contribute": "歡迎前往協助翻譯：\n${url}",
     "core.message.locale.langlist": "已支援的語言碼：${langlist}",
+    "core.message.locale.prompt": "當前使用的語言：${lang}",
     "core.message.locale.reload.failed": "以下字串重新加載失敗：\n${detail}",
     "core.message.locale.set.invalid": "語言碼無效。",
-    "core.message.locale.set.prompt": "可使用「${prefix}locale <語言碼>」設定語言。",
+    "core.message.locale.set.prompt": "使用「${prefix}locale <語言碼>」設定語言。",
     "core.message.module.disable.already": "失敗：「${module}」模組已停用。",
     "core.message.module.disable.base": "失敗：「${module}」為基礎模組，無法停用。",
     "core.message.module.disable.not_found": "失敗：「${module}」模組不存在。",
@@ -129,7 +132,7 @@
     "core.message.module.enable.not_found": "失敗：「${module}」模組不存在。",
     "core.message.module.enable.qqchannel_global.success": "成功：為所有文字頻道啟用「${module}」模組。",
     "core.message.module.enable.success": "成功：啟用模組「${module}」。",
-    "core.message.module.list.prompt": "使用「${prefix}help <對應模組名>」檢視詳細資訊。",
+    "core.message.module.list.prompt": "使用「${prefix}help <模組名稱>」檢視詳細資訊。",
     "core.message.module.load.failed": "載入模組失敗。",
     "core.message.module.load.not_found": "失敗：此模組已載入或不存在。",
     "core.message.module.load.success": "成功載入模組：${module}",
@@ -149,16 +152,16 @@
     "core.message.module.unload.not_found": "失敗：此模組不存在。",
     "core.message.module.unload.success": "成功移除模組：${module}",
     "core.message.module.unload.unavailable.confirm": "此模組已由於發生錯誤而無法載入，此動作將會在下一次重新啟動後完全停用此模組，是否繼續？",
-    "core.message.module.unsupported_language": "提示：「${module}」模組不支援目前語言，此模組可能不會正確展示目前語言下的內容。",
+    "core.message.module.unsupported_language": "提示：「${module}」模組不支援當前語言，此模組可能不會正確展示當前語言下的內容。",
     "core.message.mute.disable": "成功取消靜音。",
     "core.message.mute.enable": "成功靜音。",
-    "core.message.petal": "目前使用者 ${sender} 的花瓣數為 ${petal}。",
+    "core.message.petal": "當前使用者 ${sender} 的花瓣數為 ${petal}。",
     "core.message.petal.clear": "已清除使用者 ${sender} 的所有花瓣。",
     "core.message.petal.clear.self": "已清除所有花瓣。",
-    "core.message.petal.modify": "已為使用者 ${sender} 修改了 ${add_petal} 花瓣，目前花瓣數為：${petal}。",
-    "core.message.petal.modify.self": "已修改 ${add_petal} 花瓣，目前花瓣數為：${petal}。",
-    "core.message.petal.self": "目前你的花瓣數為 ${petal}。",
-    "core.message.ping.detail": "系統啟動時間：${system_boot_time}\n機器人已運行：${bot_running_time}\nPython 版本：${python_version}\nWebrender 狀態：${web_render_status}\n用戶端名稱：${client_name}\n自啟動以來已處理指令：${command_parsed}\n處理器型號：${cpu_brand}\n目前處理器使用率：${cpu_usage}%\n物理記憶體：${ram}M 使用率：${ram_percent}%\nSwap 記憶體：${swap}M 使用率：${swap_percent}%\n磁碟容量：${disk_space}G/${disk_space_total}G",
+    "core.message.petal.modify": "已為使用者 ${sender} 修改了 ${add_petal} 花瓣，當前花瓣數為：${petal}。",
+    "core.message.petal.modify.self": "已修改 ${add_petal} 花瓣，當前花瓣數為：${petal}。",
+    "core.message.petal.self": "當前你的花瓣數為 ${petal}。",
+    "core.message.ping.detail": "系統啟動時間：${system_boot_time}\n機器人已運行：${bot_running_time}\nPython 版本：${python_version}\nWebrender 狀態：${web_render_status}\n用戶端名稱：${client_name}\n自啟動以來已處理指令：${command_parsed}\n處理器型號：${cpu_brand}\n當前處理器使用率：${cpu_usage}%\n物理記憶體：${ram}M 使用率：${ram_percent}%\nSwap 記憶體：${swap}M 使用率：${swap_percent}%\n磁碟容量：${disk_space}G/${disk_space_total}G",
     "core.message.ping.simple": "機器人已運行：${bot_running_time}\n伺服器磁碟使用率：${disk_percent}%",
     "core.message.prefix.add.already": "此指令前綴已存在於自訂前綴列表。",
     "core.message.prefix.add.success": "已新增自訂指令前綴：${prefix}",
@@ -168,7 +171,7 @@
     "core.message.prefix.remove.not_found": "此指令前綴不存在於自訂前綴列表。",
     "core.message.prefix.remove.success": "已移除自訂指令前綴：${prefix}",
     "core.message.prefix.reset": "已重設自訂指令前綴列表。",
-    "core.message.purge.empty": "目前快取目錄為空。",
+    "core.message.purge.empty": "當前快取目錄為空。",
     "core.message.purge.success": "成功清除快取。",
     "core.message.raise": "例外測試。",
     "core.message.restart.prompt": "由於機器人正在重新啟動，此次執行指令的後續操作已被取消。\n請稍後重新執行指令，對此帶來的不便，我們深感抱歉。",
@@ -197,7 +200,7 @@
     "core.message.setup.typing.enable": "已啟用輸入提醒。",
     "core.message.update.binary_mode": "机器人正在使用二进制部署，无法更新。请手动下载最新版本的二进制文件进行替换。",
     "core.message.update.failed": "嘗試從 GitHub 取得更新失敗。",
-    "core.message.version": "目前機器人版本號：${commit}",
+    "core.message.version": "當前機器人版本號：${commit}",
     "core.message.version.unknown": "無法取得版本號。此實例可能沒有使用 Git 進行部署。",
     "core.message.whoami": "你的 ID 是：${sender}\n此對話的 ID 是：${target}",
     "core.message.whoami.admin": "（你擁有此對話的管理員權限）",
