+++ modules/wiki/locales/zh_tw.json
@@ -6,14 +6,14 @@
     "wiki.help.headers.add": "新增自訂請求標頭。",
     "wiki.help.headers.remove": "刪除自訂請求標頭。",
     "wiki.help.headers.reset": "重設自訂請求標頭。",
-    "wiki.help.headers.show": "列出目前設定的請求標頭。",
+    "wiki.help.headers.show": "列出當前設定的請求標頭。",
     "wiki.help.id": "依據頁面 ID 查詢一個 Wiki 頁面。",
     "wiki.help.iw.add": "新增自訂 Interwiki。",
     "wiki.help.iw.get": "取得設定的 Interwiki 對應的 API 位址。",
-    "wiki.help.iw.list": "列出目前設定的 Interwiki。",
+    "wiki.help.iw.list": "列出當前設定的 Interwiki。",
     "wiki.help.iw.remove": "刪除自訂 Interwiki。",
     "wiki.help.newbie": "取得預設 Wiki 的最新使用者日誌。",
-    "wiki.help.option.l": "查找本頁面的對應語言版本，若無結果則返回目前語言。",
+    "wiki.help.option.l": "查找本頁面的對應語言版本，若無結果則返回當前語言。",
     "wiki.help.prefix.reset": "重設自訂前綴。",
     "wiki.help.prefix.set": "設定自訂前綴。",
     "wiki.help.rc": "取得預設 Wiki 的近期變更。",
@@ -41,7 +41,7 @@
     "wiki.message.headers.add.failed": "更新請求標頭失敗，请校對輸入。",
     "wiki.message.headers.add.success": "已更新請求時所使用的請求標頭：${headers}",
     "wiki.message.headers.reset.success": "已重設請求時所使用的請求標頭。",
-    "wiki.message.headers.show": "目前設定了以下請求標頭：\n${headers}\n如需自訂，請使用「${prefix}wiki headers add」。 \n示例：\n${prefix}wiki headers add {\"accept-language\": \"zh-TW,zh;q=0.9,en;q=0.8,en-GB;q=0.7,en-US;q=0.6\"}",
+    "wiki.message.headers.show": "當前設定了以下請求標頭：\n${headers}\n如需自訂，請使用「${prefix}wiki headers add」。 \n示例：\n${prefix}wiki headers add {\"accept-language\": \"zh-TW,zh;q=0.9,en;q=0.8,en-GB;q=0.7,en-US;q=0.6\"}",
     "wiki.message.id.invalid": "頁面 ID 必須為數字。",
     "wiki.message.id.not_found": "未找到頁面 ID 為 ${id} 的頁面。",
     "wiki.message.invalid.blocked": "失敗：${name} 處於黑名單中。",
@@ -52,8 +52,8 @@
     "wiki.message.iw.add.success": "已新增自訂 Interwiki：\n${iw} -> ${name}",
     "wiki.message.iw.get.not_found": "未找到 Interwiki：${iw}",
     "wiki.message.iw.list": "使用 「${prefix}wiki iw get <Interwiki>」可以取得 Interwiki 對應的連結。",
-    "wiki.message.iw.list.legacy": "目前設定了以下 Interwiki：",
-    "wiki.message.iw.list.none": "目前沒有設定任何 Interwiki，可使用 「${prefix}wiki iw add」新增。",
+    "wiki.message.iw.list.legacy": "當前設定了以下 Interwiki：",
+    "wiki.message.iw.list.none": "當前沒有設定任何 Interwiki，可使用 「${prefix}wiki iw add」新增。",
     "wiki.message.iw.list.prompt": "此處展示的是機器人設定的自訂 Interwiki，如需查看起始 Wiki 的 Interwiki，請見：${url}",
     "wiki.message.iw.remove.success": "已刪除自訂 Interwiki：${iw}",
     "wiki.message.magic_word": "提示：機器人暫不支援魔術字。",
@@ -167,7 +167,7 @@
     "wiki.message.wiki_audit.add.success": "成功加入${list_name}：${api}",
     "wiki.message.wiki_audit.list.allowlist": "現有白名單：",
     "wiki.message.wiki_audit.list.blocklist": "現有黑名單：",
-    "wiki.message.wiki_audit.list.none": "目前名單為空。",
+    "wiki.message.wiki_audit.list.none": "當前名單為空。",
     "wiki.message.wiki_audit.list.table.header.apilink": "API 連結",
     "wiki.message.wiki_audit.list.table.header.date": "新增日期",
     "wiki.message.wiki_audit.list_name.allowlist": "白名單",
@@ -180,7 +180,7 @@
     "wiki.message.wiki_audit.remove.failed.not_api": "失敗：所給的連結不是網站 API（連結應以「api.php」結尾）。",
     "wiki.message.wiki_audit.remove.failed.other": "失敗：${api} 不存在於白名單中，此 Wiki 的白名單可能來自其他同一域名的 Wiki。",
     "wiki.message.wiki_audit.remove.success": "成功從${list_name}刪除：${api}",
-    "wiki.message.wiki_audit.untrust": "注意：此 Wiki 目前沒有加入機器人的白名單列表中，查詢此 Wiki 時將會對返回內容進行限制。",
+    "wiki.message.wiki_audit.untrust": "注意：此 Wiki 當前沒有加入機器人的白名單列表中，查詢此 Wiki 時將會對返回內容進行限制。",
     "wiki.message.wiki_audit.untrust.address": "如需取消限制，請在此處申請白名單：\n${url}",
     "wiki.message.wiki_bot.login.failed": "登入失敗：\n${detail}",
     "wiki.message.wiki_bot.login.success": "登入成功。",
