+++ pyproject.toml
@@ -66,6 +66,7 @@ emoji = "^2.12.1"
 langconv = "0.3.0"
 ff3 = "^1.0.2"
 orjson = "^3.10.9"
+jinja2 = "^3.1.4"
 
 
 [tool.poetry.group.dev.dependencies]
